finished refactoring scss and module information, next is to fix front-end css code layout
