+++ client/src/components/Footer/Footer.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../../../../css/components/Footer.css';
+import styles from './Footer.module.scss';
 
 export default function Footer(props) {
   return (
