+++ client/src/components/Projects/HR/Projects/Frontier/FrontierRealty.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Button, Modal, ModalHeader, ModalBody, ModalFooter } from 'reactstrap';
-import styles from '../../../../../../../css/components/HrProjects.css';
-import icons from '../../../../../../../css/components/Technical.css';
+import styles from '../../HrProjects.module.scss';
+import icons from '../../../../About/Technical/styles/Technical.module.scss';
 
 export default class FrontierRealty extends Component {
   constructor(props) {
