+++ client/src/components/Projects/HR/Projects/SDC/Sdc.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { Button, Modal, ModalHeader, ModalBody, ModalFooter } from 'reactstrap';
-import styles from '../../../../../../../css/components/HrProjects.css';
+import styles from '../../HrProjects.module.scss';
+import icons from '../../../../About/Technical/styles/Technical.module.scss';
 
 export default class Sdc extends Component {
   constructor(props) {
