--- css/components/AboutMe.css
@@ -1,165 +1,0 @@
-body h1 {
-  font-family: Arial, Helvetica, sans-serif;
-  font-family: 'Allerta Stencil', sans-serif;
-  line-height: 1.5em;
-  letter-spacing: 0.4px;
-  font-size: 3em;
-  font-size: 1.3em; }
-
-body h2 {
-  font-family: Arial, Helvetica, sans-serif;
-  font-family: 'Allerta Stencil', sans-serif;
-  line-height: 1.5em;
-  letter-spacing: 0.4px;
-  font-size: 3em;
-  margin: 2rem 0 2rem 0;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  align-items: center; }
-
-body h5 {
-  font-family: Arial, Helvetica, sans-serif;
-  font-family: 'Allerta Stencil', sans-serif;
-  line-height: 1.5em;
-  letter-spacing: 0.4px;
-  font-size: 3em;
-  font-size: 1.3em; }
-
-body h6 {
-  font-family: Arial, Helvetica, sans-serif;
-  font-family: 'Allerta Stencil', sans-serif;
-  line-height: 1.5em;
-  letter-spacing: 0.4px;
-  font-size: 3em;
-  font-size: 1.3em; }
-
-body p {
-  font-family: Arial, Helvetica, sans-serif;
-  font-family: 'Oswald', sans-serif;
-  line-height: 1.5em;
-  letter-spacing: 0.4px;
-  font-size: 1.4em; }
-
-.containerMargin {
-  margin: 10rem 0 6rem 0; }
-
-.tableNoMargin {
-  margin-bottom: 0; }
-
-.title {
-  z-index: -10;
-  /* The height needs to be set to a fixed value for the effect to work.
-   * 100vh is the full height of the viewport. */
-  height: 100vh;
-  /* The scaling of the images would add a horizontal scrollbar, so disable x overflow. */
-  overflow-x: hidden;
-  /* Enable scrolling on the page. */
-  overflow-y: auto;
-  /* Set the perspective to 2px. This is essentailly the simulated distance from the viewport to transformed objects.*/
-  perspective: 4.5px; }
-
-.parallaxHider {
-  background-color: #e0e3da; }
-
-.headings2 {
-  margin-top: 1.5rem;
-  text-align: center;
-  margin-bottom: 1.5rem; }
-
-.background1 {
-  background-image: url("https://images.pexels.com/photos/39811/pexels-photo-39811.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=750&w=1260");
-  background-repeat: no-repeat; }
-
-.background2 {
-  background-image: url("https://images.pexels.com/photos/39811/pexels-photo-39811.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=750&w=1260"); }
-
-.primaryBodyContainer {
-  width: 100%; }
-
-.aboutMe {
-  position: relative;
-  z-index: 10;
-  bottom: 13rem; }
-
-.details {
-  margin-top: -10rem;
-  border-top: thin solid #6eb1b6;
-  border-bottom: thin solid #6eb1b6; }
-
-.popUpModal {
-  z-index: 1;
-  bottom: 14rem;
-  overflow: hidden;
-  position: absolute;
-  left: 28rem; }
-
-.mainWrapper {
-  margin: auto; }
-
-footer {
-  margin: 0;
-  background-color: rgba(187, 6, 6, 0.8); }
-
-.barStyle {
-  width: 65%;
-  height: 0.2rem;
-  border-radius: 10px;
-  background-color: rgba(187, 6, 6, 0.8); }
-
-@font-face {
-  font-family: 'Edmonsans-Bold';
-  src: url("https://s3.amazonaws.com/personal-site-tim/fonts/Edmondsans-Bold.otf"); }
-
-@font-face {
-  font-family: 'Edmonsans-Medium';
-  src: url("https://s3.amazonaws.com/personal-site-tim/fonts/Edmondsans-Medium.otf"); }
-
-@font-face {
-  font-family: 'Edmonsans-Reg';
-  src: url("https://s3.amazonaws.com/personal-site-tim/fonts/Edmondsans-Regular.otf"); }
-
-.noBorder {
-  border-bottom: none;
-  background-color: transparent;
-  border: none; }
-
-.paragraphWidth {
-  margin-bottom: 6rem;
-  /* border: 1px solid rgba(0, 0, 0, 0.125); */
-  text-align: justify; }
-
-.aboutMeBg {
-  background-color: #fff;
-  margin: auto;
-  padding: 0;
-  font-family: Simplifica; }
-
-.aboutMeFooter {
-  background-color: rgba(187, 6, 6, 0.8);
-  margin: none;
-  height: 5rem; }
-
-.shadow {
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-  background-position: center;
-  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1); }
-  .shadow:hover {
-    box-shadow: 0 6px 28px rgba(0, 0, 0, 0.25), 0 6px 10px rgba(0, 0, 0, 0.22); }
-
-.aboutMeImg {
-  width: 19rem;
-  margin-top: 0.5rem;
-  max-width: fit-content; }
-
-.jumboTronBg {
-  background-color: #fff;
-  margin-bottom: 0;
-  letter-spacing: 0.4px; }
-  .jumboTronBg h1 {
-    font-family: 'Dancing Script', cursive;
-    font-size: 5em; }
-  .jumboTronBg h4 {
-    font-family: 'Allerta Stencil', sans-serif; }
-  .jumboTronBg table {
-    font-family: 'Oswald', sans-serif; }
