--- css/components/Headline.css
@@ -1,15 +1,0 @@
-.title {
-  display: flex;
-  justify-content: center;
-  align-content: center;
-  font-family: SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New', monospace;
-  font-weight: 900;
-  height: 858px;
-  position: absolute;
-  top: 30%;
-  left: 40%;
-  background-color: transparent; }
-
-.heading1 {
-  position: absolute;
-  top: 50%; }
