--- css/components/HrProjects.css
@@ -1,252 +1,0 @@
-body h1 {
-  font-family: Arial, Helvetica, sans-serif;
-  font-family: 'Allerta Stencil', sans-serif;
-  line-height: 1.5em;
-  letter-spacing: 0.4px;
-  font-size: 3em;
-  font-size: 1.3em; }
-
-body h2 {
-  font-family: Arial, Helvetica, sans-serif;
-  font-family: 'Allerta Stencil', sans-serif;
-  line-height: 1.5em;
-  letter-spacing: 0.4px;
-  font-size: 3em;
-  margin: 2rem 0 2rem 0;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  align-items: center; }
-
-body h5 {
-  font-family: Arial, Helvetica, sans-serif;
-  font-family: 'Allerta Stencil', sans-serif;
-  line-height: 1.5em;
-  letter-spacing: 0.4px;
-  font-size: 3em;
-  font-size: 1.3em; }
-
-body h6 {
-  font-family: Arial, Helvetica, sans-serif;
-  font-family: 'Allerta Stencil', sans-serif;
-  line-height: 1.5em;
-  letter-spacing: 0.4px;
-  font-size: 3em;
-  font-size: 1.3em; }
-
-body p {
-  font-family: Arial, Helvetica, sans-serif;
-  font-family: 'Oswald', sans-serif;
-  line-height: 1.5em;
-  letter-spacing: 0.4px;
-  font-size: 1.4em; }
-
-.containerMargin {
-  margin: 10rem 0 6rem 0; }
-
-.tableNoMargin {
-  margin-bottom: 0; }
-
-.title {
-  z-index: -10;
-  /* The height needs to be set to a fixed value for the effect to work.
-   * 100vh is the full height of the viewport. */
-  height: 100vh;
-  /* The scaling of the images would add a horizontal scrollbar, so disable x overflow. */
-  overflow-x: hidden;
-  /* Enable scrolling on the page. */
-  overflow-y: auto;
-  /* Set the perspective to 2px. This is essentailly the simulated distance from the viewport to transformed objects.*/
-  perspective: 4.5px; }
-
-.parallaxHider {
-  background-color: #e0e3da; }
-
-.headings2 {
-  margin-top: 1.5rem;
-  text-align: center;
-  margin-bottom: 1.5rem; }
-
-.background1 {
-  background-image: url("https://images.pexels.com/photos/39811/pexels-photo-39811.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=750&w=1260");
-  background-repeat: no-repeat; }
-
-.background2 {
-  background-image: url("https://images.pexels.com/photos/39811/pexels-photo-39811.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=750&w=1260"); }
-
-.primaryBodyContainer {
-  width: 100%; }
-
-.aboutMe {
-  position: relative;
-  z-index: 10;
-  bottom: 13rem; }
-
-.details {
-  margin-top: -10rem;
-  border-top: thin solid #6eb1b6;
-  border-bottom: thin solid #6eb1b6; }
-
-.popUpModal {
-  z-index: 1;
-  bottom: 14rem;
-  overflow: hidden;
-  position: absolute;
-  left: 28rem; }
-
-.mainWrapper {
-  margin: auto; }
-
-footer {
-  margin: 0;
-  background-color: rgba(187, 6, 6, 0.8); }
-
-.barStyle {
-  width: 65%;
-  height: 0.2rem;
-  border-radius: 10px;
-  background-color: rgba(187, 6, 6, 0.8); }
-
-.projectFont {
-  font-family: Edmonsans-Medium;
-  color: #dee2e6;
-  margin: 0; }
-
-.cardMarginAndBg {
-  background-color: transparent; }
-
-.shoedidasImgBg {
-  height: 18rem;
-  margin: 2rem 0 2rem 0;
-  background: url("https://s3.amazonaws.com/personal-site-tim/projects/adidas.jpeg");
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.12), 0 2px 8px rgba(0, 0, 0, 0.24);
-  background-size: cover;
-  background-repeat: no-repeat;
-  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
-  cursor: pointer; }
-  .shoedidasImgBg:hover {
-    box-shadow: 0 6px 28px rgba(0, 0, 0, 0.25), 0 6px 10px rgba(0, 0, 0, 0.22); }
-
-.frontierImgBg {
-  height: 18rem;
-  margin: 2rem 0 2rem 0;
-  background: url("https://s3.amazonaws.com/personal-site-tim/projects/home_new.jpeg");
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.12), 0 2px 8px rgba(0, 0, 0, 0.24);
-  background-size: cover;
-  background-repeat: no-repeat;
-  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
-  cursor: pointer; }
-  .frontierImgBg:hover {
-    box-shadow: 0 6px 28px rgba(0, 0, 0, 0.25), 0 6px 10px rgba(0, 0, 0, 0.22); }
-
-.p {
-  font-family: Arial, Helvetica, sans-serif;
-  font-family: 'Oswald', sans-serif;
-  line-height: 1.5em;
-  letter-spacing: 0.4px;
-  font-size: 18px; }
-
-.popUpModalHidden {
-  visibility: hidden; }
-
-.materialBtn {
-  border: 1px solid #000;
-  cursor: pointer;
-  background-color: transparent;
-  transition: all 0.8s cubic-bezier(0.25, 0.8, 0.25, 1); }
-  .materialBtn:hover {
-    background: linear-gradient(to bottom, transparent 0%, rgba(240, 239, 239, 0.363) 100%); }
-  .materialBtn:active {
-    background-color: rgba(240, 239, 239, 0.253) 100%;
-    background-size: 100%;
-    transition: background 0s;
-    border: 1px solid #000; }
-  .materialBtn:focus {
-    border: 1px solid #000;
-    outline: none; }
-
-.techStackIcons {
-  width: 2.5rem;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.12), 0 2px 0px rgba(0, 0, 0, 0.24);
-  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
-  border-radius: 3px; }
-
-.modalCardTechStack {
-  border: 1px solid #dee2e6;
-  background-color: #e0e3da;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1); }
-
-.modalImg {
-  margin-bottom: 2rem;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-  transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1); }
-
-.showDetails {
-  top: 5rem;
-  border: 0;
-  opacity: 0;
-  box-sizing: border-box;
-  transition: all 200ms;
-  text-align: initial;
-  transition: all 0.5s; }
-  .isActive .showDetails {
-    top: 5rem;
-    border: 0;
-    font-size: 24px;
-    opacity: 0;
-    box-sizing: border-box;
-    transition: all 200ms;
-    text-align: initial;
-    transition: all 0.5s; }
-    :hover > .isActive .showDetails {
-      top: 5rem;
-      border: 0;
-      font-size: 24px;
-      text-align: initial;
-      opacity: 1; }
-
-.githubBg {
-  background: url("https://s3.amazonaws.com/personal-site-tim/github_svg.svg");
-  background-size: cover;
-  height: 33px;
-  border: 1px solid #000;
-  opacity: 0.5; }
-
-.viewMore {
-  background: url("https://s3.amazonaws.com/personal-site-tim/eye-regular.svg");
-  background-size: auto;
-  height: 33px;
-  border: 1px solid #000;
-  opacity: 0.5; }
-
-.descriptionCardPosition {
-  position: relative; }
-
-.shoedidasImg {
-  width: 0; }
-
-.headingBg {
-  color: #fff;
-  transition: all 0.5s; }
-  .isActive .headingBg {
-    background-color: #dee2e6;
-    opacity: 0.5; }
-    :hover > .isActive .headingBg {
-      opacity: 1;
-      background-color: #dee2e6;
-      -webkit-transition: background-color 300ms linear;
-      -ms-transition: background-color 300ms linear;
-      transition: background-color 300ms linear; }
-
-.projectOverlayBg {
-  opacity: 1;
-  transition: all 0.5s;
-  height: 100%; }
-  .projectOverlayBg:hover {
-    background-color: rgba(0, 0, 0, 0.6);
-    -webkit-transition: background-color 300ms linear;
-    -ms-transition: background-color 300ms linear;
-    transition: background-color 300ms linear;
-    z-index: 1;
-    height: 100%; }
