--- scss/index.scss
@@ -1,1 +1,0 @@
-@import 'custom';
